fix: split apps inventory surfaces
